+++ src/commands/hardis/package/version/create.ts
@@ -5,6 +5,7 @@ import { AnyJson } from '@salesforce/ts-types';
 import * as c from 'chalk';
 import { execSfdxJson } from '../../../../common/utils';
 import { prompts } from '../../../../common/utils/prompts';
+import { getConfig, setConfig } from '../../../../config';
 
 // Initialize Messages with the current plugin directory
 Messages.importMessagesDirectory(__dirname);
@@ -42,23 +43,36 @@ export default class PackageVersionCreate extends SfdxCommand {
 
     public async run(): Promise<AnyJson> {
         const debugMode = this.flags.debug || false;
-
+        const config = await getConfig("project");
+        // List project packages
         const packageDirectories = this.project.getUniquePackageDirectories();
         console.error(JSON.stringify(packageDirectories));
         const packageResponse = await prompts([
             {
                 type: 'select', 
-                name: 'package',
-                message: c.cyanBright(`Please select a package`),
+                name: 'packageSelected',
+                message: c.cyanBright(`Please select a package (this is not a drill, it will create an official new version !)`),
                 choices: packageDirectories.map(packageDirectory => {
-                        return { title: packageDirectory.package, value: packageDirectory}
+                        return { title: (packageDirectory.package || packageDirectory.path), value: packageDirectory.name}
                     })
+            },
+            {
+                type: 'text', 
+                name: 'packageInstallationKey',
+                message: c.cyanBright(`Please input an installation password (or let empty)`),
+                initial: config.defaultPackageInstallationKey || ''
             }
         ]);
+        // Manage user response
+        const pckgDirectory = packageDirectories.filter(pckgDirectory => pckgDirectory.name === packageResponse.packageSelected)[0];
+        if (config.defaultPackageInstallationKey !== packageResponse.packageInstallationKey) {
+            await setConfig("project", {defaultPackageInstallationKey: packageResponse.packageInstallationKey})
+        }
 
+        // Create package version
         const createCommand = 'sfdx force:package:version:create' +
-            ` -p ${packageResponse.package.package}` +
-            ((packageResponse.package.packageInstallationKey ? ` --installationkey ${packageResponse.package.packageInstallationKey}` : ' --installationkeybypass')) +
+            ` -p "${pckgDirectory.package}"` +
+            ((packageResponse.packageInstallationKey ? ` --installationkey "${packageResponse.packageInstallationKey}"` : ' --installationkeybypass')) +
             ' -w 60';
         const createResult = await execSfdxJson(createCommand, this, { fail: true, output: true, debug: debugMode });
         const latestVersion = createResult.result.SubscriberPackageVersionId;