+++ src/common/utils/xmlUtils.ts
@@ -1,5 +1,6 @@
 // XML Utils functions
 import * as fs from 'fs-extra';
+import * as path from 'path';
 import * as xml2js from 'xml2js';
 
 export async function parseXmlFile(xmlFile: string) {
@@ -11,5 +12,6 @@ export async function parseXmlFile(xmlFile: string) {
 export async function writeXmlFile(xmlFile: string, xmlObject: any) {
     const builder = new xml2js.Builder();
     const updatedFileContent = builder.buildObject(xmlObject);
+    await fs.ensureDir(path.dirname(xmlFile));
     await fs.writeFile(xmlFile,updatedFileContent);
 }